chore: set python version to 3.13
Python 3.13 is available in Ubuntu 25.10, 25.04, 24.10. We know
some of us work with Ubuntu, so this simplifies things a bit.

We're going to use `uv` anyway, which allows to use any python
version very easily. So, it should not be a problem to use
the proper Python version under other systems.

Also, 3.14 is the latest release, while 3.13 has been around
for about one year at this point and seems a reasonable choice.