fix(data): include sample counts in JSON files (#32)
The queries were already gathering the sample counts. However, this
information was not exposed to JSON. We expose two sample counts, since
download and upload are distinct ndt7 subtests.

While there, make sure the download query correctly only checks non-null
loss rates. Should not be an issue because the loss rate should be
there. However, it seems more robust to be more precise with respect to
this matter.

Also, obviously, re-run the pipeline to get updated JSON files. (I am
increasingly skeptical that we should cache the JSON files since the
expensive operation is querying BigQuery and the flapping is caused by
re-running BigQuery, so I view quite favorably caching the BigQuery
output with parquet and then using pandas to generate the data we need.)

While there, let us take note that rerunning the pipeline causes
flapping in some JSON files. This is expected due to BigQuery being a
distributed database and `APPROX_QUANTILE`. However, it is worth keeping
in mind.