feat(cache)!: use dataset granularity (#61)
This diff modifies the cache to use the dataset granularity.

With this change implemented, the cache now builds a suitable dataset
name for m-lab without knowing details about how it will actually be
represented on disk.

This makes the cache more independent of the disk layout, which now is
mostly a concern of the pipeline code.

BREAKING CHANGE: `get_cache_entry` now requires the dataset granularity
rather than a string.